+++ packages/components/src/components/autocomplete/autocomplete.tsx
@@ -34,7 +34,9 @@ import { guid } from "../../utils/guid";
 import { useT9n } from "../../controllers/useT9n";
 import type { Input } from "../input/input";
 import type { AutocompleteItem } from "../autocomplete-item/autocomplete-item";
+import { isAutocompleteItem } from "../autocomplete-item/resources";
 import type { AutocompleteItemGroup } from "../autocomplete-item-group/autocomplete-item-group";
+import { isAutocompleteItemGroup } from "../autocomplete-item-group/resources";
 import type { Label } from "../label/label";
 import { InternalLabel } from "../functional/InternalLabel";
 import { Validation } from "../functional/Validation";
@@ -302,7 +304,7 @@ export class Autocomplete
    */
   @property({ readOnly: true }) validity!: ValidityState;
 
-  /** Specifies the selected `autocomplete-item`. When the component resides in a form, the `value` is submitted with the form. */
+  /** Specifies the value of the selected `autocomplete-item`. When the component resides in a form, the `value` is submitted with the form. */
   @property() value = "";
 
   //#endregion
@@ -423,6 +425,10 @@ export class Autocomplete
       this.openHandler();
     }
 
+    if (changes.has("value") && this.hasUpdated) {
+      this.selectedItemsHandler();
+    }
+
     if (
       changes.has("flipPlacements") ||
       (changes.has("overlayPositioning") &&
@@ -492,6 +498,10 @@ export class Autocomplete
     }
   }
 
+  private selectedItemsHandler(): void {
+    this.items.forEach((item) => (item.selected = item.value === this.value));
+  }
+
   private openHandler(): void {
     if (this.disabled) {
       this.open = false;
@@ -517,6 +527,7 @@ export class Autocomplete
 
   private async handleAutocompleteItemSelect(event: Event): Promise<void> {
     this.value = (event.target as AutocompleteItem["el"]).value;
+    this.selectedItemsHandler();
     this.emitChange();
     await this.setFocus();
     this.open = false;
@@ -579,6 +590,10 @@ export class Autocomplete
   private updateItems(): void {
     let activeDescendant = "";
 
+    if (this.value) {
+      this.selectedItemsHandler();
+    }
+
     this.items.forEach((item) => {
       item.scale = this.scale;
       item.inputValueMatchPattern = this.inputValueMatchPattern;
@@ -624,23 +639,15 @@ export class Autocomplete
           this.defaultSlotRef.value,
           groupItemSelector,
         )
-      : Array.from(this.el.children).filter((child): child is AutocompleteItemGroup["el"] =>
-          child.matches(groupItemSelector),
-        );
+      : Array.from(this.el.children).filter(isAutocompleteItemGroup);
 
     const rootItems = this.defaultSlotRef.value
       ? getSlotAssignedElements<AutocompleteItem["el"]>(this.defaultSlotRef.value, itemSelector)
-      : Array.from(this.el.children).filter((child): child is AutocompleteItem["el"] =>
-          child.matches(itemSelector),
-        );
+      : Array.from(this.el.children).filter(isAutocompleteItem);
 
     const groupedItems = groups.flatMap((group) => group.items ?? []);
     const items = Array.from(
-      new Set<AutocompleteItem["el"]>(
-        [...rootItems, ...groupedItems].filter(
-          (item): item is AutocompleteItem["el"] => !!item && item.matches(itemSelector),
-        ),
-      ),
+      new Set<AutocompleteItem["el"]>([...rootItems, ...groupedItems].filter(isAutocompleteItem)),
     );
 
     this.groups = groups;
@@ -678,8 +685,7 @@ export class Autocomplete
         break;
       case "Enter":
         if (open && activeItem) {
-          this.value = activeItem.value;
-          activeItem.toggleSelection();
+          activeItem.requestSelection();
           this.open = false;
           event.preventDefault();
         } else if (!event.defaultPrevented && this.formSupport.active) {